requirements.txt
fastapi[standard]
python_dotenv
bitsandbytes
transformers
datasets
requests
torch
peft
